+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/metamodel/attribute/AbstractMethodListAttribute.java
@@ -24,12 +24,15 @@ import com.blazebit.persistence.view.impl.metamodel.AbstractMethodPluralAttribut
 import com.blazebit.persistence.view.impl.metamodel.EmbeddableOwner;
 import com.blazebit.persistence.view.impl.metamodel.ManagedViewTypeImplementor;
 import com.blazebit.persistence.view.impl.metamodel.MetamodelBuildingContext;
+import com.blazebit.persistence.view.impl.metamodel.MetamodelUtils;
 import com.blazebit.persistence.view.impl.metamodel.MethodAttributeMapping;
 import com.blazebit.persistence.view.impl.objectbuilder.CollectionInstantiatorAccumulator;
 import com.blazebit.persistence.view.impl.objectbuilder.ContainerAccumulator;
 import com.blazebit.persistence.view.metamodel.MethodListAttribute;
 import com.blazebit.persistence.view.metamodel.Type;
+import com.blazebit.persistence.view.spi.EntityViewAttributeMapping;
 
+import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 
@@ -44,6 +47,11 @@ public abstract class AbstractMethodListAttribute<X, Y> extends AbstractMethodPl
     private final Expression indexMappingExpression;
     private final boolean isIndexed;
     private final boolean forcedUnique;
+    private final boolean elementCollectionSorted;
+    private final boolean elementCollectionOrdered;
+    private final boolean elementCollectionForcedUnique;
+    private final Class<Comparator<Object>> elementCollectionComparatorClass;
+    private final Comparator<Object> elementCollectionComparator;
     private final CollectionInstantiatorAccumulator collectionInstantiatorAccumulator;
     
     public AbstractMethodListAttribute(ManagedViewTypeImplementor<X> viewType, MethodAttributeMapping mapping, MetamodelBuildingContext context, int attributeIndex, int dirtyStateIndex, EmbeddableOwner embeddableMapping) {
@@ -60,9 +68,14 @@ public abstract class AbstractMethodListAttribute<X, Y> extends AbstractMethodPl
         this.indexMapping = indexMapping;
         this.indexMappingExpression = createSimpleExpression(indexMapping, mapping, context, ExpressionLocation.MAPPING_INDEX);
         this.forcedUnique = mapping.isForceUniqueness() || determineForcedUnique(context);
+        this.elementCollectionSorted = mapping.getElementCollectionBehavior() == EntityViewAttributeMapping.ElementCollectionBehavior.SORTED || getElementCollectionType() == ElementCollectionType.SORTED_SET;
+        this.elementCollectionOrdered = mapping.getElementCollectionBehavior() == EntityViewAttributeMapping.ElementCollectionBehavior.ORDERED || getElementCollectionType() == ElementCollectionType.LIST;
+        this.elementCollectionComparatorClass = (Class<Comparator<Object>>) mapping.getElementCollectionComparatorClass();
+        this.elementCollectionComparator = MetamodelUtils.getComparator(elementCollectionComparatorClass);
+        this.elementCollectionForcedUnique = mapping.isElementCollectionForceUniqueness();
         this.collectionInstantiatorAccumulator = new CollectionInstantiatorAccumulator(
                 createCollectionInstantiator(context, createCollectionFactory(context), isIndexed(), isSorted(), isOrdered(), getComparator()),
-                createValueContainerAccumulator(),
+                createValueContainerAccumulator(elementCollectionComparator),
                 !isCorrelated()
         );
     }
@@ -105,6 +118,31 @@ public abstract class AbstractMethodListAttribute<X, Y> extends AbstractMethodPl
     @Override
     public boolean isForcedUnique() {
         return forcedUnique;
+    }
+
+    @Override
+    public boolean isElementCollectionOrdered() {
+        return elementCollectionOrdered;
+    }
+
+    @Override
+    public boolean isElementCollectionSorted() {
+        return elementCollectionSorted;
+    }
+
+    @Override
+    public boolean isElementCollectionForcedUnique() {
+        return elementCollectionForcedUnique;
+    }
+
+    @Override
+    public Comparator<?> getElementCollectionComparator() {
+        return elementCollectionComparator;
+    }
+
+    @Override
+    public Class<Comparator<?>> getElementCollectionComparatorClass() {
+        return (Class<Comparator<?>>) (Class<?>) elementCollectionComparatorClass;
     }
 
     @Override