+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/metamodel/attribute/AbstractParameterMapAttribute.java
@@ -25,13 +25,16 @@ import com.blazebit.persistence.view.impl.metamodel.EmbeddableOwner;
 import com.blazebit.persistence.view.impl.metamodel.ManagedViewTypeImplementor;
 import com.blazebit.persistence.view.impl.metamodel.MappingConstructorImpl;
 import com.blazebit.persistence.view.impl.metamodel.MetamodelBuildingContext;
+import com.blazebit.persistence.view.impl.metamodel.MetamodelUtils;
 import com.blazebit.persistence.view.impl.metamodel.ParameterAttributeMapping;
 import com.blazebit.persistence.view.impl.objectbuilder.ContainerAccumulator;
 import com.blazebit.persistence.view.impl.objectbuilder.MapInstantiatorAccumulator;
 import com.blazebit.persistence.view.metamodel.ManagedViewType;
 import com.blazebit.persistence.view.metamodel.MapAttribute;
 import com.blazebit.persistence.view.metamodel.Type;
+import com.blazebit.persistence.view.spi.EntityViewAttributeMapping;
 
+import java.util.Comparator;
 import java.util.Map;
 
 /**
@@ -47,6 +50,11 @@ public abstract class AbstractParameterMapAttribute<X, K, V> extends AbstractPar
     private final Type<K> keyType;
     private final Map<ManagedViewType<? extends K>, String> keyInheritanceSubtypes;
     private final boolean forcedUnique;
+    private final boolean elementCollectionSorted;
+    private final boolean elementCollectionOrdered;
+    private final boolean elementCollectionForcedUnique;
+    private final Class<Comparator<Object>> elementCollectionComparatorClass;
+    private final Comparator<Object> elementCollectionComparator;
     private final MapInstantiatorAccumulator mapInstantiatorAccumulator;
 
     @SuppressWarnings("unchecked")
@@ -62,9 +70,14 @@ public abstract class AbstractParameterMapAttribute<X, K, V> extends AbstractPar
         this.keyType = (Type<K>) mapping.getKeyType(context, embeddableMapping);
         this.keyInheritanceSubtypes = (Map<ManagedViewType<? extends K>, String>) (Map<?, ?>) mapping.getKeyInheritanceSubtypes(context, embeddableMapping);
         this.forcedUnique = mapping.isForceUniqueness() || determineForcedUnique(context);
+        this.elementCollectionSorted = mapping.getElementCollectionBehavior() == EntityViewAttributeMapping.ElementCollectionBehavior.SORTED || getElementCollectionType() == ElementCollectionType.SORTED_SET;
+        this.elementCollectionOrdered = mapping.getElementCollectionBehavior() == EntityViewAttributeMapping.ElementCollectionBehavior.ORDERED || getElementCollectionType() == ElementCollectionType.LIST;
+        this.elementCollectionComparatorClass = (Class<Comparator<Object>>) mapping.getElementCollectionComparatorClass();
+        this.elementCollectionComparator = MetamodelUtils.getComparator(elementCollectionComparatorClass);
+        this.elementCollectionForcedUnique = mapping.isElementCollectionForceUniqueness();
         this.mapInstantiatorAccumulator = new MapInstantiatorAccumulator(
                 createMapInstantiator(context, null, isSorted(), isOrdered(), getComparator()),
-                createValueContainerAccumulator(),
+                createValueContainerAccumulator(elementCollectionComparator),
                 !isCorrelated()
         );
     }
@@ -72,6 +85,31 @@ public abstract class AbstractParameterMapAttribute<X, K, V> extends AbstractPar
     @Override
     public boolean isForcedUnique() {
         return forcedUnique;
+    }
+
+    @Override
+    public boolean isElementCollectionOrdered() {
+        return elementCollectionOrdered;
+    }
+
+    @Override
+    public boolean isElementCollectionSorted() {
+        return elementCollectionSorted;
+    }
+
+    @Override
+    public boolean isElementCollectionForcedUnique() {
+        return elementCollectionForcedUnique;
+    }
+
+    @Override
+    public Comparator<?> getElementCollectionComparator() {
+        return elementCollectionComparator;
+    }
+
+    @Override
+    public Class<Comparator<?>> getElementCollectionComparatorClass() {
+        return (Class<Comparator<?>>) (Class<?>) elementCollectionComparatorClass;
     }
 
     @Override