+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/objectbuilder/SimpleCollectionAccumulator.java
@@ -18,7 +18,13 @@ package com.blazebit.persistence.view.impl.objectbuilder;
 
 import com.blazebit.persistence.view.impl.collection.PluralObjectFactory;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.Set;
 
 /**
  *
@@ -28,9 +34,13 @@ import java.util.Collection;
 public class SimpleCollectionAccumulator implements ContainerAccumulator<Collection<Object>>  {
 
     private final PluralObjectFactory<Collection<Object>> pluralObjectFactory;
+    private final boolean forceUnique;
+    private final Comparator<Object> comparator;
 
-    public SimpleCollectionAccumulator(PluralObjectFactory<? extends Collection<?>> pluralObjectFactory) {
+    public SimpleCollectionAccumulator(PluralObjectFactory<? extends Collection<?>> pluralObjectFactory, boolean forceUnique, Comparator<Object> comparator) {
         this.pluralObjectFactory = (PluralObjectFactory<Collection<Object>>) pluralObjectFactory;
+        this.forceUnique = forceUnique;
+        this.comparator = comparator;
     }
 
     @Override
@@ -50,10 +60,25 @@ public class SimpleCollectionAccumulator implements ContainerAccumulator<Collect
 
     @Override
     public boolean requiresPostConstruct() {
-        return false;
+        return forceUnique || comparator != null;
     }
 
     @Override
     public void postConstruct(Collection<Object> collection) {
+        ArrayList<Object> list = (ArrayList<Object>) collection;
+        if (forceUnique) {
+            Set<Object> set = new HashSet<>(list.size());
+            Iterator<Object> iter = list.iterator();
+
+            while (iter.hasNext()) {
+                Object o = iter.next();
+                if (!set.add(o)) {
+                    iter.remove();
+                }
+            }
+        }
+        if (comparator != null) {
+            Collections.sort(list, comparator);
+        }
     }
 }